Fixed visual studio compilation error / warnings
git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13098 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/fdc/DirAsDSK.cc
@@ -618,7 +618,7 @@ void DirAsDSK::importHostFile(DirIndex dirIndex, FileOperations::Stat& fst)
 	}
 
 	// Write (possibly truncated) file size.
-	msxDir(dirIndex).size = hostSize - remainingSize;
+	msxDir(dirIndex).size = uint32_t(hostSize - remainingSize);
 
 	// TODO in case of an error (disk image full, or host file read error),
 	// wouldn't it be better to remove the (half imported) msx file again?
